Move artifact writing functionality to separate file.
As we are adding more functionalities to the library we need
to isolate code specific to different functionalities. This one
is moving artifact writing part to write.go file.

Changelog: None
